packages/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
tensorflow==2.13.0
tensorflow-hub==0.12.0
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
numpy==1.24.3
python-multipart==0.0.6
